refactor(interpreter): split formulas functions and matches
